# trail-server/target/classes/api-spec.yaml
swagger: '2.0'
info:
  version: 0.1.0
  title: Trailer API
  description: An API to demonstrate Swagger and Spring Boot
host: 'localhost:8081'
basePath: /api
schemes:
  - http
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
  /users:
    post:
      security:
        - Bearer: []
      description: create a user
      operationId: createUser
      consumes:
        - application/json
      parameters:
        - name: user
          in: body
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        201:
          description: created
          schema:
            type: object
  /users/{email}:
    get:
      security:
        - Bearer: []
      description: ""
      operationId: getUserByID
      produces:
        - "application/json"
      parameters:
        - name: email
          in: path
          description: "The name that needs to be fetched. Use user1 for testing. "
          required: true
          type: string
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid username supplied"
        404:
          description: "User not found"
    put:
      security:
        - Bearer: []
      description: ""
      operationId: updateUserByID
      produces:
        - "application/json"
      parameters:
        - name: email
          in: path
          description: "name that need to be updated"
          required: true
          type: string
        - in: body
          name: userUpdate
          description: "Updated user object"
          required: true
          schema:
            $ref: "#/definitions/UserUpdate"
      responses:
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
    delete:
      security:
        - Bearer: []
      operationId: deleteUserByID
      produces:
        - "application/json"
      parameters:
        - name: email
          in: path
          description: "The name that needs to be deleted"
          required: true
          type: string
      responses:
        400:
          description: "Invalid username supplied"
        404:
          description: "User not found"
...
definitions:
  UserUpdate:
    type: object
    properties:
      firstname:
        type: string
      lastname:
        type: string
      date:
        type: string
  User:
    allOf:
      - $ref: "#/definitions/UserUpdate"
      - type: object
        properties:
          email:
            type: string
...